feat: query logging classes
