--- src/main/java/com/sprint/mission/discodeit/config/LogConfiguration.java
@@ -1,14 +1,0 @@
-package com.sprint.mission.discodeit.config;
-
-import com.p6spy.engine.spy.P6SpyOptions;
-import jakarta.annotation.PostConstruct;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class LogConfiguration {
-
-  @PostConstruct
-  public void setLogMessageFormat() {
-    P6SpyOptions.getActiveInstance().setLogMessageFormat(MyP6spyFormattingStrategy.class.getName());
-  }
-}
