+++ src/main/java/com/sprint/mission/discodeit/logging/LogConfiguration.java
@@ -1,0 +1,14 @@
+package com.sprint.mission.discodeit.logging;
+
+import com.p6spy.engine.spy.P6SpyOptions;
+import jakarta.annotation.PostConstruct;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class LogConfiguration {
+
+  @PostConstruct
+  public void setLogMessageFormat() {
+    P6SpyOptions.getActiveInstance().setLogMessageFormat(MyP6spyFormattingStrategy.class.getName());
+  }
+}
